[ERD-68] Factor out inspection panel. Add custom expansion and filtering behaviour.
